Minor fixes with function signatures and Vec3D

+++ src/vec3d.rs
@@ -1,11 +1,11 @@
 pub struct Vec3D {
-  x: u32,
-  y: u32,
-  z: u32
+  x: f64,
+  y: f64,
+  z: f64
 }
 
 impl Vec3D {
-  pub fn new(x: u32, y: u32, z:u32) -> Vec3D {
+  pub fn new(x: f64, y: f64, z:f64) -> Vec3D {
       let new_vec3d = Vec3D {
           x,
           y,
